backend api routes secured. We can now only interract with products only when we are logged in

+++ frontend/src/components/ProductDetails.js
@@ -1,14 +1,22 @@
 import { useProductsContext } from "../hooks/useProductsContext"
+import { useAuthContext } from "../hooks/useAuthContext"
 
 // date fns
 import formatDistanceToNow from 'date-fns/formatDistanceToNow'
 
 const ProductDetails = ({ product }) => {
   const { dispatch } = useProductsContext()
+  const { user } = useAuthContext()
 
   const handleClick = async() => {
+    if (!user) {
+      return
+    }
     const response = await fetch('/api/products/' + product._id, {
-      method: 'DELETE'
+      method: 'DELETE',
+      headers: {
+        'Authorization': `Bearer ${user.token}`
+      }
     })
     const json = await response.json()
